Refactor rest core
* remove unused methods
* move mgmt resource dedicated classes to mgmt-resource
* clean up code
