+++ hawkbit-rest/hawkbit-mgmt-api/src/main/java/org/eclipse/hawkbit/mgmt/rest/api/SortDirection.java
@@ -1,0 +1,25 @@
+/**
+ * Copyright (c) 2015 Bosch Software Innovations GmbH and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.hawkbit.mgmt.rest.api;
+
+/**
+ * A definition of possible sorting direction.
+ */
+public enum SortDirection {
+
+    /**
+     * Ascending.
+     */
+    ASC,
+    /**
+     * Descending.
+     */
+    DESC;
+}
